requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
django-debug-toolbar
drf-nested-routers
Pillow
redis
Celery
requests